+++ packages/oui-calendar/tests/index.js
@@ -1,0 +1,7 @@
+import "@ovh-ui/common/test-utils";
+
+loadTests(require.context("../src/", true, /.*((\.spec)|(index))$/));
+
+function loadTests (context) {
+    context.keys().forEach(context);
+}