+++ packages/oui-select/src/index.spec.js
@@ -19,6 +19,8 @@ describe("ouiSelect", () => {
 
     const getContainer = element => element[0].querySelector(".ui-select-container");
     const getDropdownButton = element => element[0].querySelector(".ui-select-match");
+    const getMultipleDropdownButton = element => element[0].querySelector(".ui-select-match-container");
+    const getMultipleMatchItem = element => element[0].querySelectorAll(".ui-select-match-item");
     const getDropdown = element => element[0].querySelector(".ui-select-choices-content");
     const getFocusser = element => element[0].querySelector(".ui-select-focusser");
     const getItemsGroups = element => element[0].querySelectorAll(".ui-select-choices-group");
@@ -38,8 +40,7 @@ describe("ouiSelect", () => {
                     title="${title}"
                     placeholder="${placeholder}"
                     items="$ctrl.countries"
-                    match="name">
-                    <span ng-bind="$item.name"></span>
+                    match="country.name">
                 </oui-select>`, {
                 countries: data
             });
@@ -57,8 +58,7 @@ describe("ouiSelect", () => {
                     title="Select a country"
                     placeholder="Select a country..."
                     items="$ctrl.countries"
-                    match="name">
-                    <span ng-bind="$item.name"></span>
+                    match="country.name">
                 </oui-select>`, {
                 countries: data
             });
@@ -84,8 +84,8 @@ describe("ouiSelect", () => {
                         title="Select a country"
                         placeholder="Select a country..."
                         items="$ctrl.countries"
-                        match="name">
-                        <span ng-bind="$item.name"></span>
+                        match="country.name">
+                        <span ng-bind="$item.country.name"></span>
                     </oui-select>
                     <button class="outside-button">Outside</button>
                 </div>`, {
@@ -111,8 +111,8 @@ describe("ouiSelect", () => {
                         title="Select a country"
                         placeholder="Select a country..."
                         items="$ctrl.countries"
-                        match="name">
-                        <span ng-bind="$item.name"></span>
+                        match="country.name">
+                        <span ng-bind="$item.country.name"></span>
                     </oui-select>`, {
                     countries: data
                 });
@@ -140,6 +140,73 @@ describe("ouiSelect", () => {
             });
         });
 
+        describe("Multiple select", () => {
+            it("should not close dropdown when an item is selected", () => {
+                const element = TestUtils.compileTemplate(`
+                    <oui-select name="country"
+                        model="$ctrl.country"
+                        title="Select a country"
+                        placeholder="Select a country..."
+                        items="$ctrl.countries"
+                        match="country.name"
+                        multiple>
+                    </oui-select>`, {
+                    countries: data
+                });
+
+                const $triggerButton = angular.element(getMultipleDropdownButton(element));
+
+                expect($triggerButton.attr("aria-expanded")).toBe("false");
+
+                // Open the dropdown
+                $triggerButton.triggerHandler("click");
+
+                // Select 5th element and check if it's highlighted.
+                const $itemButton = angular.element(getDropdownItem(element, 4)); // eslint-disable-line no-magic-numbers
+                $itemButton.triggerHandler("click");
+
+                // The dropdown should stay opened.
+                expect($triggerButton.attr("aria-expanded")).toBe("true");
+            });
+
+            it("should remove item selected", () => {
+                const element = TestUtils.compileTemplate(`
+                    <oui-select name="country"
+                        model="$ctrl.country"
+                        title="Select a country"
+                        placeholder="Select a country..."
+                        items="$ctrl.countries"
+                        match="country.name"
+                        multiple>
+                    </oui-select>`, {
+                    countries: data
+                });
+
+                const $triggerButton = angular.element(getMultipleDropdownButton(element));
+
+                expect($triggerButton.attr("aria-expanded")).toBe("false");
+
+                // Open the dropdown
+                $triggerButton.triggerHandler("click");
+
+                // Select 5th element and check if it's highlighted.
+                let $itemButton = angular.element(getDropdownItem(element, 4)); // eslint-disable-line no-magic-numbers
+                $itemButton.triggerHandler("click");
+
+                $itemButton = angular.element(getDropdownItem(element, 4)); // eslint-disable-line no-magic-numbers
+                $itemButton.triggerHandler("click");
+
+                // The dropdown should stay opened.
+                let matchItems = getMultipleMatchItem(element);
+                expect(matchItems.length).toBe(2); // eslint-disable-line no-magic-numbers
+
+                angular.element(matchItems[0]).triggerHandler("click");
+                matchItems = getMultipleMatchItem(element);
+
+                expect(matchItems.length).toBe(1); // eslint-disable-line no-magic-numbers
+            });
+        });
+
         describe("Not grouped", () => {
             it("should display all the choices (objectArray)", () => {
                 const element = TestUtils.compileTemplate(`
@@ -148,8 +215,7 @@ describe("ouiSelect", () => {
                         title="Select a country"
                         placeholder="Select a country..."
                         items="$ctrl.countries"
-                        match="name">
-                        <span ng-bind="$item.name"></span>
+                        match="country.name">
                     </oui-select>`, {
                     countries: data
                 });
@@ -159,8 +225,8 @@ describe("ouiSelect", () => {
                 $triggerButton.triggerHandler("click");
 
                 expect(getDropdownItems(element).length).toEqual(data.length);
-                expect(angular.element(getDropdownItem(element, 0)).text()).toContain(data[0].name);
-                expect(angular.element(getDropdownItem(element, data.length - 1)).text()).toContain(data[data.length - 1].name);
+                expect(angular.element(getDropdownItem(element, 0)).text()).toContain(data[0].country.name);
+                expect(angular.element(getDropdownItem(element, data.length - 1)).text()).toContain(data[data.length - 1].country.name);
                 expect(getItemsGroups(element).length).toEqual(1);
             });
 
@@ -171,7 +237,6 @@ describe("ouiSelect", () => {
                         title="Select a country"
                         model="$ctrl.country"
                         items="$ctrl.array">
-                        <span ng-bind="$item"></span>
                     </oui-select>`, {
                     array: stringArray
                 });
@@ -189,16 +254,15 @@ describe("ouiSelect", () => {
 
         describe("Grouped", () => {
             it("should display all the choices", () => {
-                const groupByFirstLetter = (item) => item.name.substr(0, 1).toUpperCase();
+                const groupByFirstLetter = (item) => item.country.name.substr(0, 1).toUpperCase();
                 const element = TestUtils.compileTemplate(`
                     <oui-select name="country"
                         model="$ctrl.country"
                         title="Select a country"
                         placeholder="Select a country..."
                         items="$ctrl.countries"
-                        match="name"
+                        match="country.name"
                         group-by="$ctrl.groupByFirstLetter">
-                        <span ng-bind="$item.name"></span>
                     </oui-select>`, {
                     countries: data,
                     groupByFirstLetter
@@ -228,9 +292,8 @@ describe("ouiSelect", () => {
                         title="Select a country"
                         placeholder="Select a country..."
                         items="$ctrl.countries"
-                        match="name"
+                        match="country.name"
                         on-blur="$ctrl.onBlur()">
-                        <span ng-bind="$item.name"></span>
                     </oui-select>`, {
                     onBlur
                 });
@@ -238,6 +301,10 @@ describe("ouiSelect", () => {
                 $timeout.flush();
 
                 angular.element(getFocusser(element)).triggerHandler("blur");
+
+                // Need to flush again for the callback
+                $timeout.flush();
+
                 expect(onBlur).toHaveBeenCalled();
             });
         });
@@ -251,9 +318,8 @@ describe("ouiSelect", () => {
                         title="Select a country"
                         placeholder="Select a country..."
                         items="$ctrl.countries"
-                        match="name"
+                        match="country.name"
                         on-focus="$ctrl.onFocus()">
-                        <span ng-bind="$item.name"></span>
                     </oui-select>`, {
                     onFocus
                 });
@@ -261,6 +327,10 @@ describe("ouiSelect", () => {
                 $timeout.flush();
 
                 angular.element(getFocusser(element)).triggerHandler("focus");
+
+                // Need to flush again for the callback
+                $timeout.flush();
+
                 expect(onFocus).toHaveBeenCalled();
             });
         });
@@ -274,9 +344,8 @@ describe("ouiSelect", () => {
                         title="Select a country"
                         placeholder="Select a country..."
                         items="$ctrl.countries"
-                        match="name"
+                        match="country.name"
                         on-change="$ctrl.onChange(modelValue)">
-                        <span ng-bind="$item.name"></span>
                     </oui-select>`, {
                     countries: data,
                     onChange
@@ -302,16 +371,15 @@ describe("ouiSelect", () => {
 
         describe("Disable options", () => {
             it("should disable corresponding items", () => {
-                const disableCountry = (item) => item.name === data[3].name;
+                const disableCountry = (item) => item.country.name === data[3].country.name;
                 const element = TestUtils.compileTemplate(`
                     <oui-select name="country"
-                                model="$ctrl.country"
-                                title="Select a country"
-                                placeholder="Select a country..."
-                                items="$ctrl.countries"
-                                disable-items="$ctrl.disableCountry($item)"
-                                match="name">
-                        <span ng-bind="$item.name"></span>
+                        model="$ctrl.country"
+                        title="Select a country"
+                        placeholder="Select a country..."
+                        items="$ctrl.countries"
+                        disable-items="$ctrl.disableCountry($item)"
+                        match="country.name">
                     </oui-select>`, {
                     countries: data,
                     disableCountry
@@ -329,13 +397,12 @@ describe("ouiSelect", () => {
                 const disableCountry = (item) => item.code === "";
                 const element = TestUtils.compileTemplate(`
                     <oui-select name="country"
-                                model="$ctrl.country"
-                                title="Select a country"
-                                placeholder="Select a country..."
-                                items="$ctrl.countries"
-                                disable-items="$ctrl.disableCountry($item)"
-                                match="name">
-                        <span ng-bind="$item.name"></span>
+                        model="$ctrl.country"
+                        title="Select a country"
+                        placeholder="Select a country..."
+                        items="$ctrl.countries"
+                        disable-items="$ctrl.disableCountry($item)"
+                        match="country.name">
                     </oui-select>`, {
                     countries,
                     disableCountry