+++ packages/oui-slideshow/src/slideshow.controller.js
@@ -62,14 +62,12 @@ export default class {
         });
     }
 
-    $destroy () {
+    $onDestroy () {
         this.closeOnBoarding();
     }
 
     closeOnBoarding () {
-        if (this.onDismiss) {
-            this.onDismiss(event);
-        }
+        this.onDismiss();
 
         this.$document.off("keydown", evt => this._triggerKeyHandler(evt));
         this.$scope.$broadcast("oui:modalOnBoarding:afterClose", this.id);